setDiscussionPreferences mutation
